DPO train: cast pixel_values to model dtype (fixes float32/bfloat16 mismatch)

+++ train_model/dpo/train.py
@@ -195,7 +195,7 @@ def train(cfg: dict) -> None:
     for epoch in range(epochs):
         for step, batch in enumerate(loader):
             batch = {k: v.to(device) for k, v in batch.items()}
-            pixel_values = batch["pixel_values"]
+            pixel_values = batch["pixel_values"].to(dtype)
             image_flags = batch["image_flags"]
 
             # -- Reference forward: adapter disabled, no grads -------------
